requirements.txt
psutil
argparse
filetype
ffmpeg-python